refactor: split `utils/core.py` with `utils/server.py`
