+++ _quarto.yml
@@ -78,8 +78,10 @@ website:
                 section: "Utilities"
                 contents:
 
-                - href: "docs/reference/clim_recal.utils.core.qmd"
-                  text: "core"
+                  - href: "docs/reference/clim_recal.utils.core.qmd"
+                    text: "core"
+                  - href: "docs/reference/clim_recal.utils.server.qmd"
+                    text: "server"
       - text: "Docker"
         href: "docs/docker-configurations.qmd"
       - text: "Contributing"
@@ -111,3 +113,4 @@ quartodoc:
         - clim_recal.debiasing.debias_wrapper
         - clim_recal.utils
         - clim_recal.utils.core
+        - clim_recal.utils.server
